+++ effective-gear/common/src/main/java/com/chimericdream/effectivegear/EffectiveGearMod.java
@@ -1,5 +1,6 @@
 package com.chimericdream.effectivegear;
 
+import com.chimericdream.effectivegear.block.EGBlocks;
 import com.chimericdream.lib.registries.ModRegistryHelper;
 import com.google.common.base.Suppliers;
 import dev.architectury.registry.registries.RegistrarManager;
@@ -16,6 +17,8 @@ public final class EffectiveGearMod {
 
     public static void init() {
         MANAGER = Suppliers.memoize(() -> RegistrarManager.get(ModInfo.MOD_ID));
+
+        EGBlocks.init();
 
         REGISTRY_HELPER.init();
     }