+++ effective-gear/common/src/main/java/com/chimericdream/effectivegear/mixin/EG$LivingEntityMixin.java
@@ -6,6 +6,7 @@ import net.minecraft.server.level.ServerLevel;
 import net.minecraft.world.damagesource.DamageSource;
 import net.minecraft.world.damagesource.DamageTypes;
 import net.minecraft.world.entity.LivingEntity;
+import net.minecraft.world.item.equipment.trim.TrimMaterials;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.ModifyVariable;
@@ -18,11 +19,17 @@ public class EG$LivingEntityMixin {
         argsOnly = true,
         name = "damage"
     )
-    private float eg$reduceSonicBoomDamageWithFullEchoShardTrim(float damage, ServerLevel level, DamageSource source) {
+    private float eg$checkForDamageReduction(float damage, ServerLevel level, DamageSource source) {
         LivingEntity self = (LivingEntity) (Object) this;
 
+        // Negate lightning damage with full copper trim
+        if (source.is(DamageTypes.LIGHTNING_BOLT) && TrimSetUtils.isWearingFullTrim(self, TrimMaterials.COPPER)) {
+            return 0.0F;
+        }
+
+        // Reduce sonic boom damage with full echo shard trim
         if (source.is(DamageTypes.SONIC_BOOM) && TrimSetUtils.isWearingFullTrim(self, Trims.ECHO_SHARD_TRIM_ID)) {
-            return damage * 0.5F;
+            return damage * 0.25F;
         }
 
         return damage;